Changing direction and plan. The project will no longer focus on plugins, as the decentralised compiler scene will cause havoc when trying to use plugins made on linux with windows.
With that I will focus on making this project work on linux first and then adapting it for windows. This is due to complicated windows relationship with C++, and the project's need to firstly work, before I start optimizing it. This should help me with focusing on development instead of porting it back and forth.

To anyone maintaining it in the future, I'm sorry about the confusion, and hope it's not going to be too bad to maintain this mess of a codebase.

Will also try and actually write some documentation... Wish me luck.

P.S.: Will remove the .idea files later